Add transforms to FApplyOnce
Add CXform::ExfInnerJoinWithInnerSelect2DynamicBitmapIndexGetApply and
CXform::ExfInnerJoinWithInnerSelect2DynamicIndexGetApply to FApplyOnce.

Every transform job in Orca is applied to a specific group expression.
We do extract an expression tree from the memo rooted from that specific
group expression. The expression tree has to match the specified pattern
associated and defined with a specific transform. The below pattern is
for the two transforms mentioned above.

+--CLogicalInnerJoin
   |--CPatternLeaf
   |--CLogicalSelect
   |  |--CLogicalDynamicGet
   |  +--CPatternTree
   +--CPatternTree

The pattern may include a specific operator or some generic pattern
expressions which signifies different type of nodes. For example
CPatternLeaf means anything in the experssion tree matches and stop
creating further expression alternatives. CPatternTree on the other hand
requires to explore all the tree alternatives that might be generated
below that node.

If CScalarSubquery appears as one of the descendents  of a CPatternTree
the number of expressions we generate for that group expression can be
significantly large and we have to apply the specific transform that
many times. This leads to significantly long planning time.

Adding the transforms to FApplyOnce ensures that the expression is
generated just once.

Signed-off-by: Sambitesh Dash <sdash@pivotal.io>